move types to typing (#1186)
