+++ setup.py
@@ -25,6 +25,7 @@ setup(
         "httpx[socks]",
         "prompt-toolkit",
         "tiktoken>=0.3.0",
+        "openai"
     ],
     classifiers=[
         "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
@@ -33,7 +34,6 @@ setup(
         "Natural Language :: English",
         "Topic :: Software Development :: Libraries :: Python Modules",
         "Programming Language :: Python :: 3 :: Only",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
